GEODE-6334: Converting types to long

+++ geode-core/src/integrationTest/java/org/apache/geode/management/bean/stats/MemberLevelStatsJUnitTest.java
@@ -513,7 +513,7 @@ public class MemberLevelStatsJUnitTest extends MBeanStatsTestCase {
     return bridge.getNetLoadsAverageLatency();
   }
 
-  private int getTotalNetSearchCompleted() {
+  private long getTotalNetSearchCompleted() {
     return bridge.getTotalNetSearchCompleted();
   }
